[FIX] website_forum, website_profile: disable editor toolbar options
Before this commit, the inline style added with some options of the
editor toolbar when editing a textarea was removed due to the
strip_style of the content field.

Since we cannot remove this strip_style for security reasons, we can
only disable options that add inline style.

This commit closes #34288

closes odoo/odoo#62906

X-original-commit: 74e4114e20186e31343af12a673fa3e4bc4eac4a
Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>